refactor: move files related to Input, StatusIndicator, and Icon
