+++ giraffe/src/components/Icon/Icon.scss
@@ -7,12 +7,12 @@
 
 @font-face {
   font-family: 'icomoon';
-  src: url('../fonts/icomoon.eot') format('embedded-opentype');
-  src: url('../fonts/icomoon.eot') format('embedded-opentype'),
-    url('../fonts/icomoon.woff2') format('woff2'),
-    url('../fonts/icomoon.ttf') format('truetype'),
-    url('../fonts/icomoon.woff') format('woff'),
-    url('../fonts/icomoon.svg') format('svg');
+  src: url('../../fonts/icomoon.eot') format('embedded-opentype');
+  src: url('../../fonts/icomoon.eot') format('embedded-opentype'),
+    url('../../fonts/icomoon.woff2') format('woff2'),
+    url('../../fonts/icomoon.ttf') format('truetype'),
+    url('../../fonts/icomoon.woff') format('woff'),
+    url('../../fonts/icomoon.svg') format('svg');
   font-weight: normal;
   font-style: normal;
 }
