+++ giraffe/src/components/Icon/Icon.scss.d.ts
