+++ giraffe/src/components/Icon/index.ts
@@ -1,0 +1,1 @@
+export {Icon} from './Icon'
