+++ giraffe/src/components/Input/StatusIndicator.scss
