+++ giraffe/src/components/Input/index.ts
@@ -1,0 +1,1 @@
+export {Input} from './Input'
